Separate xml & json into separate pipelines.
